+++ R/function_PlotPerformanceByAttribute.R
@@ -93,7 +93,7 @@
 #' )
 #' }
 #' 
-#' @importFrom dplyr group_by sym left_join n rename select summarize n_distinct
+#' @importFrom dplyr group_by sym left_join n rename select summarize n_distinct pull
 #' @importFrom ggplot2 aes coord_cartesian coord_flip element_text geom_density geom_boxplot geom_point geom_smooth ggplot ggsave scale_color_manual scale_fill_manual scale_x_continuous
 #' scale_y_continuous theme theme_void unit waiver xlab ylab scale_x_log10 scale_y_log10
 #' @importFrom ggpubr colnames_style get_legend ggarrange ggtexttable tab_add_title tbody_style ttheme
@@ -307,7 +307,7 @@ PlotPerformanceByAttribute <- function(subass, subass.column = 2, groups = NULL,
           group_by(.data[["Group"]]) %>%
           summarize(n = n_distinct(!!sym(col))) %>%
           filter(n > 1) %>%
-          pull(Group)
+          pull(.data[["Group"]])
       } else {
         trendline_levels <- levels(plotdata$Group)
       }