+++ NEWS.md
@@ -7,6 +7,7 @@
 * Add ReadBestsims() function
 * Fix behavior of colors and drop in PlotPerformanceByAttribute()
 * Add additional geometry types to PlotParvalues()
+* Bug fixes to OutletNearObs()
 
 ## HYPEtools 1.6.7 (2026-04-08)
 *Enhancements and bug fixes*