+++ NEWS.md
@@ -9,6 +9,7 @@
 * Add additional geometry types to PlotParvalues()
 * Bug fixes to OutletNearObs()
 * Fix bug in na.color for PlotMapOutput()
+* Add CalculateRiverDistance() function
 
 ## HYPEtools 1.6.7 (2026-04-08)
 *Enhancements and bug fixes*